+++ packages/netlify-cms-widget-relation/src/__tests__/relation.spec.js
@@ -16,6 +16,22 @@ const fieldConfig = {
   valueField: 'title',
 };
 
+const deeplyNestedFieldConfig = {
+  name: 'post',
+  collection: 'posts',
+  displayFields: ['title', 'slug', 'deeply.nested.post.field'],
+  searchFields: ['deeply.nested.post.field'],
+  valueField: 'title',
+};
+
+const nestedFieldConfig = {
+  name: 'post',
+  collection: 'posts',
+  displayFields: ['title', 'slug', 'nested.field_1'],
+  searchFields: ['nested.field_1', 'nested.field_2'],
+  valueField: 'title',
+};
+
 const generateHits = length => {
   const hits = Array.from({ length }, (val, idx) => {
     const title = `Post # ${idx + 1}`;
@@ -25,6 +41,31 @@ const generateHits = length => {
 
   return [
     ...hits,
+    {
+      collection: 'posts',
+      data: {
+        title: 'Deeply nested post',
+        slug: 'post-deeply-nested',
+        deeply: {
+          nested: {
+            post: {
+              field: 'Deeply nested field',
+            },
+          },
+        },
+      },
+    },
+    {
+      collection: 'posts',
+      data: {
+        title: 'Nested post',
+        slug: 'post-nested',
+        nested: {
+          field_1: 'Nested field 1',
+          field_2: 'Nested field 2',
+        },
+      },
+    },
     {
       collection: 'posts',
       data: { title: 'YAML post', slug: 'post-yaml', body: 'Body yaml' },
@@ -51,6 +92,14 @@ class RelationController extends React.Component {
     const queryHits = generateHits(25);
     if (last(args) === 'YAML') {
       return Promise.resolve({ payload: { response: { hits: [last(queryHits)] } } });
+    } else if (last(args) === 'Nested') {
+      return Promise.resolve({
+        payload: { response: { hits: [queryHits[queryHits.length - 2]] } },
+      });
+    } else if (last(args) === 'Deeply nested') {
+      return Promise.resolve({
+        payload: { response: { hits: [queryHits[queryHits.length - 3]] } },
+      });
     }
     return Promise.resolve({ payload: { response: { hits: queryHits } } });
   });
@@ -156,6 +205,28 @@ describe('Relation widget', () => {
       expect(getByText(label)).toBeInTheDocument();
       expect(onChangeSpy).toHaveBeenCalledTimes(1);
       expect(onChangeSpy).toHaveBeenCalledWith(value, metadata);
+    });
+  });
+
+  it('should update option list based on nested search term', async () => {
+    const field = fromJS(nestedFieldConfig);
+    const { getAllByText, input } = setup({ field });
+    fireEvent.change(input, { target: { value: 'Nested' } });
+
+    await wait(() => {
+      expect(getAllByText('Nested post post-nested Nested field 1')).toHaveLength(1);
+    });
+  });
+
+  it('should update option list based on deeply nested search term', async () => {
+    const field = fromJS(deeplyNestedFieldConfig);
+    const { getAllByText, input } = setup({ field });
+    fireEvent.change(input, { target: { value: 'Deeply nested' } });
+
+    await wait(() => {
+      expect(
+        getAllByText('Deeply nested post post-deeply-nested Deeply nested field'),
+      ).toHaveLength(1);
     });
   });